pyproject.toml
[project]
name = "healthcare-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipython>=9.6.0",
    "langchain[together]>=1.0.0a10",
    "langgraph>=1.0.0a4",
    "python-dotenv>=1.1.1",
]
